Some minor changes in preparation for spray-ification. Note run/parse still a giant placeholder
