+++ src/main/scala/cromwell/Main.scala
@@ -2,12 +2,15 @@ package cromwell
 
 import java.io.File
 import java.nio.file.Paths
+import java.util.UUID
 
 import akka.actor.ActorSystem
 import cromwell.binding._
 import cromwell.binding.values.{WdlString, WdlFile, WdlValue}
 import cromwell.engine.WorkflowManagerActor
-import cromwell.engine.WorkflowManagerActor.SubmitWorkflow
+import cromwell.engine.WorkflowManagerActor.{WorkflowStatus, SubmitWorkflow}
+import cromwell.server.CromwellServer
+import cromwell.singleuser.SingleWorkflowRunner
 
 
 import scala.util.{Failure, Success, Try}
@@ -27,8 +30,8 @@ object Main extends App {
   override def main(args: Array[String]) {
     getAction(args.headOption) match {
       case Some(x) if x == Actions.parse => parseIt(args.tail)
-      case Some(x) if x == Actions.run => runIt(args.tail)
-      case Some(x) if x == Actions.server => serverIt(args.tail)
+      case Some(x) if x == Actions.run => SingleWorkflowRunner.runIt(args.tail)
+      case Some(x) if x == Actions.server => CromwellServer // Mention it so it gets loaded. Note that the actor system is also lazily initialized
       case _ => usageAndExit()
     }
   }
@@ -36,7 +39,7 @@ object Main extends App {
   /**
    * Placeholder for now
    */
-  private def printProcessedBinding(binding: WdlBinding): Unit = {
+  def printProcessedBinding(binding: WdlBinding): Unit = {
     println(s"Workflow FQN: ${binding.workflow.fullyQualifiedName}")
     println(s"Workflow Inputs: ${binding.workflow.inputs}")
     println(s"Workflow Outputs: ${binding.workflow.outputs}")
@@ -68,57 +71,6 @@ object Main extends App {
   def parseIt(args: Array[String]): Unit = {
     if (args.isEmpty) usageAndExit()
     else println(WdlBinding.getAst(new File(args(1))).toPrettyString)
-  }
-
-  /** YAP */
-  def runIt(args: Array[String]): Unit = {
-    if (args.isEmpty) usageAndExit()
-    else {
-      Try(WdlBinding.process(new File(args(1)))) match {
-        case Success(b) => printProcessedBinding(b)
-        case Failure(e) =>
-          println(e)
-          System.exit(-1)
-      }
-    }
-  }
-
-  /** YAP */
-  def serverIt(args: Array[String]): Unit = {
-    val systemName = "cromwell-system"
-    val actorSystem = ActorSystem(systemName)
-
-    actorSystem.registerOnTermination({
-      actorSystem.log.info(s"$systemName shutting down")
-    })
-
-    val workflowManagerActor = actorSystem.actorOf(WorkflowManagerActor.props(actorSystem))
-
-    // TODO: Awaiting spray stuff, and really - once that's here this whole thing should move somewhere else
-
-
-    println("Cromwell server started")
-    // FIXME: Just testing for now
-    val HelloWdl =
-      """
-        |task hello {
-        |  command {
-        |    echo "Hello ${addressee}!"
-        |  }
-        |  output {
-        |    String salutation = read_string("stdout")
-        |  }
-        |}
-        |
-        |workflow hello {
-        |  call hello
-        |}
-      """.stripMargin
-
-    val Addressee = "hello.hello.addressee"
-    val HelloInputs: Map[FullyQualifiedName, WdlValue] = Map(Addressee -> WdlString("world"))
-
-    workflowManagerActor ! SubmitWorkflow(HelloWdl, HelloInputs)
   }
 
   /**
