+++ src/main/scala/cromwell/server/WorkflowManagerSystem.scala
@@ -1,0 +1,13 @@
+package cromwell.server
+
+import akka.actor.ActorSystem
+import cromwell.engine.WorkflowManagerActor
+
+trait WorkflowManagerSystem {
+  val systemName = "cromwell-system"
+  val actorSystem = ActorSystem(systemName)
+
+  actorSystem.registerOnTermination {actorSystem.log.info(s"$systemName shutting down")}
+
+  val workflowManagerActor = actorSystem.actorOf(WorkflowManagerActor.props)
+}
